* 'new_stack' of github.com:Workiva/frugal:
  Fix tests
  Print friendly error message on compile failure
  Handle recursive frugal and thrift includes properly
  Run the pubspec generator over every pubspec (even if no frugal)

+++ compiler/thrift.go
@@ -8,7 +8,6 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/Workiva/frugal/compiler/globals"
 	"github.com/Workiva/frugal/compiler/parser"
 )
 
@@ -33,6 +32,11 @@ var thriftTypes = map[string]bool{
 
 func generateThriftIDL(dir string, frugal *parser.Frugal) (string, error) {
 	file := filepath.Join(dir, fmt.Sprintf("%s.thrift", frugal.Name))
+	if exists(file) {
+		// Trying to generate an intermediate Thrift IDL but the .thrift file
+		// already exists.
+		return "", fmt.Errorf("Thrift file already exists: %s", file)
+	}
 	f, err := os.Create(file)
 	if err != nil {
 		return "", err
@@ -71,26 +75,33 @@ func generateNamespaces(namespaces []*parser.Namespace) string {
 
 func generateIncludes(frugal *parser.Frugal) (string, error) {
 	contents := ""
+	// Recurse on includes
 	for _, incl := range frugal.Thrift.Includes {
 		include := incl.Value
-		if strings.HasSuffix(strings.ToLower(include), ".frugal") {
-			// Recurse on frugal includes
-			parsed, err := compile(filepath.Join(frugal.Dir, include))
-			if err != nil {
-				return "", err
-			}
-
-			// Lop off .frugal
-			includeBase := include[:len(include)-7]
-
-			// Lop off path
-			includeName := filepath.Base(includeBase)
-
-			frugal.ParsedIncludes[includeName] = parsed
-
-			// Replace .frugal with .thrift
-			include = includeBase + ".thrift"
+		pathAndExtension := strings.Split(strings.ToLower(include), ".")
+		if len(pathAndExtension) != 2 {
+			return "", fmt.Errorf("Bad include name: %s", include)
 		}
+		extension := pathAndExtension[1]
+		if extension != "thrift" && extension != "frugal" {
+			return "", fmt.Errorf("Bad include extension: %s", include)
+		}
+
+		parsed, err := compile(filepath.Join(frugal.Dir, include), extension == "thrift")
+		if err != nil {
+			return "", err
+		}
+
+		// Lop off extension (.frugal or .thrift)
+		includeBase := include[:len(include)-7]
+
+		// Lop off path
+		includeName := filepath.Base(includeBase)
+
+		frugal.ParsedIncludes[includeName] = parsed
+
+		// Replace .frugal with .thrift
+		include = includeBase + ".thrift"
 		contents += fmt.Sprintf("include \"%s\"\n", include)
 	}
 	contents += "\n"
@@ -342,25 +353,14 @@ func generateServices(services []*parser.Service) string {
 	return contents
 }
 
-func generateThrift(frugal *parser.Frugal, idlDir, out, gen string, dryRun bool) error {
-	// Generate intermediate Thrift IDL.
-	idlFile, err := generateThriftIDL(idlDir, frugal)
-	if err != nil {
-		return err
-	}
-	globals.IntermediateIDL = append(globals.IntermediateIDL, idlFile)
-
-	if dryRun {
-		return nil
-	}
-
+func generateThrift(frugal *parser.Frugal, idlDir, file, out, gen string) error {
 	// Generate Thrift code.
 	args := []string{}
 	if out != "" {
 		args = append(args, "-out", out)
 	}
 	args = append(args, "-gen", gen)
-	args = append(args, idlFile)
+	args = append(args, file)
 	// TODO: make thrift command configurable
 	if out, err := exec.Command("thrift", args...).CombinedOutput(); err != nil {
 		fmt.Println(string(out))
